+++ src/cards/assets/national-emblem.png
